Convert api to ts
